refactor(worker): Organise services into Crawling/ and Maintenance/ subfolders
Move Worker + WorkerState into Anichron.Worker.Crawling and
TokenCleanupService into Anichron.Worker.Maintenance; update all
dependent namespaces and using directives accordingly.
